Create dspam-trainer
This is the core dspam training file, it's not the prettiest.. but it works for me ( disclaimer, on a small server handling not much work, no idea on how it scales, but chances are if you know how to make things scale you can already write better than this script :). 

I wrote it after trying the built-in demo bash-scripts that never seemed to function properly. It is also likely that it duplicates a lot of things in dspam that I haven't been able to see working, and yet it probably does.. just not very clearly through the very opaque black box of how dspam works.

Anyhow, if it makes anything easier for you, then that's good enough for me.